part1/main/api/mini_ecommerce/requirements.txt
asgiref==3.10.0
Django==5.2
django-filter==25.2
djangorestframework==3.16.1
Markdown==3.9
sqlparse==0.5.3
